Mitigate the bias of the password function
Shuffle the charset before password generation and lower the length
limit for the password (MAX_PW_LENGTH).

math/rand/v2 is used for shuffling, but it should not significantly
impact security, as the generation itself uses cryptographically
secure RNG and the distribution of the generated elements is quite
close to random most of the time.